# app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "矛盾调解受理服务 API",
    "description": "提供身份验证和矛盾调解受理单管理的RESTful API",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "schemes": ["http"],
  "paths": {
...
    "/identity/verify": {
      "post": {
        "summary": "身份验证",
        "description": "通过身份证号验证用户是否存在",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "API令牌",
            "required": true,
            "type": "string",
            "default": "api_token_2025"
          },
          {
            "name": "body",
            "in": "body",
            "description": "请求体",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id_card_number": {
                  "type": "string",
                  "example": "330102199001011234"
                }
              },
              "required": ["id_card_number"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "验证成功或失败",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "integer",
                  "example": 1
                },
                "message": {
                  "type": "string",
                  "example": "身份证号验证通过"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id_card_number": {
                      "type": "string",
                      "example": "330102199001011234"
                    },
                    "name": {
                      "type": "string",
                      "example": "张三"
                    },
                    "contact_info": {
                      "type": "string",
                      "example": "13800138001"
                    },
                    "address": {
                      "type": "string",
                      "example": "浙江省杭州市西湖区文三路123号"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "integer",
                  "example": 0
                },
                "message": {
                  "type": "string",
                  "example": "参数错误：缺少身份证号"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "integer",
                  "example": 0
                },
                "message": {
                  "type": "string",
                  "example": "未授权：API令牌无效或已过期"
                }
              }
            }
          }
        }
      }
    },
    "/identity/status": {
      "get": {
        "summary": "查询验证状态",
        "description": "查询身份证号对应的用户信息和验证状态",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "API令牌",
            "required": true,
            "type": "string",
            "default": "api_token_2025"
          },
          {
            "name": "id_card_number",
            "in": "query",
            "description": "身份证号",
            "required": true,
            "type": "string",
            "example": "330102199001011234"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功或失败",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "integer",
                  "example": 1
                },
                "message": {
                  "type": "string",
                  "example": "查询成功"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id_card_number": {
                      "type": "string",
                      "example": "330102199001011234"
                    },
                    "name": {
                      "type": "string",
                      "example": "张三"
                    },
                    "contact_info": {
                      "type": "string",
                      "example": "13800138001"
                    },
                    "address": {
                      "type": "string",
                      "example": "浙江省杭州市西湖区文三路123号"
                    },
                    "verified": {
                      "type": "boolean",
                      "example": true
                    },
                    "verification_result": {
                      "type": "string",
                      "example": "验证通过"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "integer",
                  "example": 0
                },
                "message": {
                  "type": "string",
                  "example": "参数错误：缺少身份证号"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "integer",
                  "example": 0
                },
                "message": {
                  "type": "string",
                  "example": "未授权：API令牌无效或已过期"
                }
              }
            }
          }
        }
      }
    },
...
  }
} 